* Fix #1815
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4326 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ vocation.cpp
@@ -206,7 +206,8 @@ bool Vocations::parseVocationNode(xmlNodePtr p)
 		{
 			if(readXMLInteger(configNode, "percentAll", intValue))
 			{
-				for(int32_t i = COMBAT_FIRST; i <= COMBAT_LAST; i = std::pow(2., i))
+				for(int32_t i = COMBAT_FIRST; i <= 
+COMBAT_LAST; i = (int32_t)std::pow(2., i))
 					voc->increaseAbsorb((CombatType_t)i, intValue);
 			}
 
@@ -268,7 +269,8 @@ bool Vocations::parseVocationNode(xmlNodePtr p)
 		{
 			if(readXMLInteger(configNode, "percentAll", intValue))
 			{
-				for(int32_t i = COMBAT_FIRST; i <= COMBAT_LAST; i = std::pow(2., i))
+				for(int32_t i = COMBAT_FIRST; i <= 
+COMBAT_LAST; i = (int32_t)std::pow(2., i))
 					voc->increaseReflect(REFLECT_PERCENT, (CombatType_t)i, intValue);
 			}
 
@@ -328,7 +330,8 @@ bool Vocations::parseVocationNode(xmlNodePtr p)
 
 			if(readXMLInteger(configNode, "chanceAll", intValue))
 			{
-				for(int32_t i = COMBAT_FIRST; i <= COMBAT_LAST; i = std::pow(2., i))
+				for(int32_t i = COMBAT_FIRST; i <= 
+COMBAT_LAST; i = (int32_t)std::pow(2., i))
 					voc->increaseReflect(REFLECT_CHANCE, (CombatType_t)i, intValue);
 			}
